Move tests from lib to bare_index
This makes all tests that use the crates.io git index to be `#[ignore]`
as they are incredibly expensive to run
when using cargo 1.70.0+, in addition to being problemtic due to disk
mutation
